+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - 2026-??-?? 4.2.10
     - persist CorrectionStatus visibility
+    - add headers when copying StripChart data to clipboard
     - LibraryMatching
         - renamed assets/example_data/matching_library to default_matching_library
         - renamed EnlightenSpectra/matching_library to user_library_spectra